refactor: (#57) 도메인 폴더, 파일명을 컨벤션에 맞춰 정리
