+++ src/auth/auth.service.ts
@@ -7,7 +7,7 @@ import { UserData } from './interfaces/user-data.interface';
 import { TokenDto } from './dto/responses/token.dto';
 import { AccessTokenDto } from './dto/responses/access-token.dto';
 import { PasswordEncoderUtil } from 'src/common/utils/password-encoder.util';
-import { AUTH_DOMAIN_ERRORS, USER_ERROR_CODES } from './constants/auth.constants';
+import { AUTH_DOMAIN_ERRORS, USER_ERROR_CODES } from './constants/auth.constant';
 import { SocialLoginPayload } from './interfaces/social-login.interface';
 
 @Injectable()
