+++ src/auth/dtos/responses/access-token.dto.ts
