+++ src/auth/guards/kakao.guard.ts
@@ -1,10 +1,10 @@
 import { BadGatewayException, Injectable, UnauthorizedException } from '@nestjs/common';
 import { AuthGuard } from '@nestjs/passport';
-import { OAUTH_ERRORS_CODES } from '../constants/auth.constants';
+import { OAUTH_ERRORS_CODES } from '../constants/auth.constant';
 
 @Injectable()
-export class GoogleGuard extends AuthGuard('google') {
-  handleRequest<TUser = any>(err: (Error & { status?: number }) | null, user: TUser, _info): TUser {
+export class KakaoGuard extends AuthGuard('kakao') {
+  handleRequest<TUser = any>(err: (Error & { status?: number }) | null, user: TUser, _info) {
     if (err || !user) {
       if (err?.status && err.status >= 500) {
         throw new BadGatewayException(OAUTH_ERRORS_CODES.OAUTH_PROVIDER_ERROR);
