+++ src/auth/tokens/token.service.ts
@@ -4,7 +4,7 @@ import { JwtService } from '@nestjs/jwt';
 import { ConfigService } from '@nestjs/config';
 import { RedisService } from '../../redis/redis.service';
 import { AccessTokenDto } from '../dto/responses/access-token.dto';
-import { ACCESS_TOKEN_ERROR_CODES } from '../constants/auth.constants';
+import { ACCESS_TOKEN_ERROR_CODES } from '../constants/auth.constant';
 
 @Injectable()
 export class TokenService {
