+++ src/room/constants/room.constant.ts
