+++ src/room/dtos/requests/update-room.dto.ts
