+++ src/room/dtos/responses/deleted-room-response.dto.ts
