+++ src/room/dtos/responses/paginated-room-response.dto.ts
