+++ src/room/entities/room.entity.ts
