+++ src/room/interfaces/room.repository.interface.ts
@@ -1,5 +1,5 @@
 import { QueryRoomDto } from '../dtos/requests/query-room.dto';
-import { Room } from '../entities/rooms.entity';
+import { Room } from '../entities/room.entity';
 
 export interface IRoomsRepository {
   createRoom(room: Room): Promise<Room>;
