+++ src/user/user.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { UsersService } from './users.service';
-import { UsersRepository } from './users.repository';
+import { UsersService } from './user.service';
+import { UsersRepository } from './user.repository';
 import { PrismaModule } from 'src/prisma/prisma.module';
 
 @Module({
